+++ src/TestDb/dbo/Stored Procedures/ProcWithMaxLengthTypes.sql
@@ -1,0 +1,10 @@
+﻿CREATE PROCEDURE [dbo].[ProcWithMaxLengthTypes]
+  @nvarchar NVARCHAR(MAX),
+  @varbinary VARBINARY(MAX),
+  @varchar VARCHAR(MAX)
+AS
+
+SELECT
+  [nvarchar] = @nvarchar,
+  [varbinary] = @varbinary,
+  [varchar] = @varchar