+++ src/TestDb/dbo/Stored Procedures/ProcWithMaxLengthTypesFromTvp.sql
@@ -1,0 +1,5 @@
+﻿CREATE PROCEDURE [dbo].[ProcWithMaxLengthTypesFromTvp]
+  @tvp dbo.MaxLengthTypes READONLY
+AS
+
+SELECT * FROM @tvp