+++ src/TestDb/dbo/User Defined Types/MaxLengthTypes.sql
@@ -1,0 +1,6 @@
+﻿CREATE TYPE [dbo].[MaxLengthTypes] AS TABLE
+(
+  [nvarchar] NVARCHAR(MAX) NOT NULL,
+  [varbinary] VARBINARY(MAX) NOT NULL,
+  [varchar] VARCHAR(MAX) NOT NULL
+)