+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
 Release notes
 ==============
 
+### 0.1.14 (2021-01-12)
+
+* Fixed bug where MAX length parameters in TVP would throw exception
+
 ### 0.1.13 (2021-01-08)
 
 * The default generated file module is now public instead of internal (since internal table DTOs can’t be used in the DTO `WithParameter` overloads).